+++ src/main/java/team/marco/voucher_management_system/service/BlacklistService.java
@@ -2,8 +2,8 @@ package team.marco.voucher_management_system.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import team.marco.voucher_management_system.console_app.repository.BlacklistRepository;
 import team.marco.voucher_management_system.model.BlacklistUser;
-import team.marco.voucher_management_system.repository.BlacklistRepository;
 
 @Service
 public class BlacklistService {
