+++ src/test/java/team/marco/voucher_management_system/properties/FilePathPropertiesTest.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 import team.marco.voucher_management_system.configuration.TestPropertyConfiguration;
+import team.marco.voucher_management_system.console_app.properties.FilePathProperties;
 
 @SpringJUnitConfig(classes = TestPropertyConfiguration.class)
 @EnableConfigurationProperties(FilePathProperties.class)
